Move SLB.Identifier tests to the correct namespace
