+++ SAGESharpTests/SLB/IdentifierTests.cs
@@ -5,7 +5,7 @@ using SAGESharp.Testing;
 using System;
 using System.Reflection;
 
-namespace SAGESharp.SLB.Tests
+namespace SAGESharp.SLB
 {
     class IdentifierTests
     {
